Added possibility to override default config file from project

+++ start.sh
@@ -13,8 +13,10 @@ if [ -d /var/etc/nginx ]; then
             fi
         done
     else
-        echo "Linking default site"
-        ln -sf /etc/nginx/conf/sites/default /etc/nginx/sites-enabled/default
+        echo "Linking default site if not exist"
+        if [ ! -f "/etc/nginx/sites-enabled/default" ]; then
+            ln -sf /etc/nginx/conf/sites/default /etc/nginx/sites-enabled/default
+        fi
     fi
 fi
